% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{annotations}
\alias{annotations}
\title{annotations()
Resultados sobre la cantidad de anotaciones en bases de datos.
Compara los resultados separando las masas extras de las comunes en rSIREM.
son masas comunes las que coinciden o est치n muy pr칩ximas entre rMSI2 y rSIREM
son masas extras las no comunes de rSIREM.
Presenta 4 violines: dos para masas comunes y dos para masas extras en 30k y 60k}
\usage{
annotations(ppm = 10)
}
\description{
annotations()
Resultados sobre la cantidad de anotaciones en bases de datos.
Compara los resultados separando las masas extras de las comunes en rSIREM.
son masas comunes las que coinciden o est치n muy pr칩ximas entre rMSI2 y rSIREM
son masas extras las no comunes de rSIREM.
Presenta 4 violines: dos para masas comunes y dos para masas extras en 30k y 60k
}
