% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fitQualityPereSirem}
\alias{fitQualityPereSirem}
\title{fitQualityPereSirem
Aisla aquellos picos de rMSI2 y rSirem que comparten al mismo pico de alta resolución como más próximo
Para ellos, se obtienen las desviaciones respecto al pico de alta resolución}
\usage{
fitQualityPereSirem(
  reference,
  testPere,
  testSirem,
  refMinMag = 1e-06,
  testMinMag = 1e-06,
  minMass = 0,
  maxMass = 0,
  histo = FALSE
)
}
\arguments{
\item{reference}{-> gaussians info from rGetGaussians() (high resolution)}

\item{testPere}{-> list from  rMSI2::LoadPeakMatrix() from rMSI2::processWizard()}

\item{testSirem}{-> gaussians info from rGetGaussians() (low resolution)}

\item{refMinMag}{-> mínima magnitud de picos para su consideración (high resolution)}

\item{testMinMag}{-> mínima magnitud de picos para su consideración (low resolution)}

\item{minMass}{-> high mass to analyze (Da)}
}
\value{
a matrix with columns: "mzrMSI2", "mzRef", "ppm", "maxDev", "mzSirem", "mzRef", "ppm", "maxDev"
mzrMSI2 -> mz from rMSI2 sample
mzRef   -> mz from high resolution sample
ppm     -> deviation from rMSI2 to high mz
maxDev -> '1' if greater than ppm of a scan; 2 if greater than 1.5 ppm of a scan
mzSirem -> mz from Sirem sample
mzRef   -> mz from high resolution sample
ppm     -> deviation from SIrem to high mz
maxDev -> '1' if greater than ppm of a scan; 2 if greater than 1.5 ppm of a scan
}
\description{
fitQualityPereSirem
Aisla aquellos picos de rMSI2 y rSirem que comparten al mismo pico de alta resolución como más próximo
Para ellos, se obtienen las desviaciones respecto al pico de alta resolución
}
