% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/test_old.R
\name{fitQualitySirem}
\alias{fitQualitySirem}
\title{fitQualitySirem
Presents an image with the centroids of two data set into a mass range,
The parameters come from rGetGaussians() and rMSI2::LoadPeakMatrix()}
\usage{
fitQualitySirem(refSiremPeak, testSiremPeak, minMass = 0, maxMass = 0)

fitQualitySirem(refSiremPeak, testSiremPeak, minMass = 0, maxMass = 0)
}
\arguments{
\item{refSiremPeak}{-> gaussians mean info from rGetGaussians() (high resolution)}

\item{testSiremPeak}{-> gaussians mean info from rGetGaussians() (low resolution)}

\item{minMass}{-> high mass to analyze (Da)
siremPeak$gaussians:
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.
MSI2peak$mass: mass vector
MSI2peak$intensity:
columns -> concentration for each mass.
rows    -> concentration for each pixel.}
}
\value{
-1 if KO; 0 if OK

-1 if KO; 0 if OK
}
\description{
fitQualitySirem
Presents an image with the centroids of two data set into a mass range,
The parameters come from rGetGaussians() and rMSI2::LoadPeakMatrix()

fitQualitySirem
Presents an image with the centroids of two data set into a mass range,
The parameters come from rGetGaussians() and rMSI2::LoadPeakMatrix()
}
