% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fitQualitySiremDeconv}
\alias{fitQualitySiremDeconv}
\title{fitQualitySiremDeconv
Identifica el pico de alta resolución más próximo al pico de baja resolución e indica la desviación
Lo hace exclusivamente sobre los picos deconvolucionados. Para info de Sirem}
\usage{
fitQualitySiremDeconv(
  reference,
  testSirem,
  testGauss,
  refMinMag = 1e-06,
  testMinMag = 1e-06,
  minMass = 0,
  maxMass = 0,
  histo = FALSE
)
}
\arguments{
\item{reference}{-> gaussians info from rGetGaussians() (high resolution)}

\item{testSirem}{-> sirem     info from rGetGaussians() (low resolution)}

\item{testGauss}{-> gaussians info from rGetGaussians() (low resolution)}

\item{refMinMag}{-> mínima magnitud de picos para su consideración (high resolution)}

\item{testMinMag}{-> mínima magnitud de picos para su consideración (low resolution)}

\item{minMass}{-> high mass to analyze (Da)}
}
\value{
a matrix with columns: "mzTest", "mzRef", "ppm", "maxDev", "repe"
mzTest -> mz from low  resolution sample
mzRef  -> mz from high resolution sample
ppm    -> deviation from low to high mz
maxDev -> '1' if greater than ppm of a scan
repe   -> '1' if mzRef is the same for two consecutives mzTest
}
\description{
fitQualitySiremDeconv
Identifica el pico de alta resolución más próximo al pico de baja resolución e indica la desviación
Lo hace exclusivamente sobre los picos deconvolucionados. Para info de Sirem
}
