% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_old.R
\name{gaussPeaksDeconvolved}
\alias{gaussPeaksDeconvolved}
\title{gaussPeaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados luego de ser
reajustados como gausianas}
\usage{
gaussPeaksDeconvolved(peaks, gaussInfo)
}
\arguments{
\item{peaks}{-> matrix returned from siremPeaksDeconvolved()}

\item{gaussInfo}{-> list returned from rGetGaussians()}
}
\value{
a matrix: rows=peaks deconvolved under a peak of magnitude;
col\link{1}to col\link{4} =mass of each gaussian
}
\description{
gaussPeaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados luego de ser
reajustados como gausianas
}
