% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{getUpeakGaussFromMass}
\alias{getUpeakGaussFromMass}
\title{getUpeakGaussFromMass
retorna un array con las masas de sirem contenidas en el mismo pico compuesto que la masa dada}
\usage{
getUpeakGaussFromMass(siremPeaks, gaussInfo, mass)
}
\arguments{
\item{siremPeaks}{-> sirem peaks, from rGetSiremPeaks()}

\item{mass}{-> masa}
}
\value{
array con las masas asociadas
}
\description{
getUpeakGaussFromMass
retorna un array con las masas de sirem contenidas en el mismo pico compuesto que la masa dada
}
