% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{getUpeakSiremMassesFromUpeak}
\alias{getUpeakSiremMassesFromUpeak}
\title{getUpeakSiremMassesFromUpeak
retorna un array con las masas de sirem asociadas a un pico de magnitud}
\usage{
getUpeakSiremMassesFromUpeak(siremPeaks, uPeak)
}
\arguments{
\item{siremPeaks}{-> sirem peaks info, from rGetSiremPeaks()}

\item{uPeak}{-> united peaks}
}
\value{
array con las masas asociadas al pico dado
}
\description{
getUpeakSiremMassesFromUpeak
retorna un array con las masas de sirem asociadas a un pico de magnitud
}
