% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{histoJC}
\alias{histoJC}
\title{histo2
Histograma adaptado para visualizar los hits en base de datos de aduptos de Jorde Capellades
bin size=1; 0>= x <=15; x.column=4}
\usage{
histoJC(x, title)
}
\arguments{
\item{x}{-> datos (read.csv("Jordi_Capellades_file"))}

\item{title}{-> título principal}
}
\value{
nothing
}
\description{
histo2
Histograma adaptado para visualizar los hits en base de datos de aduptos de Jorde Capellades
bin size=1; 0>= x <=15; x.column=4
}
