% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R, R/test_old.R
\name{nearestValue}
\alias{nearestValue}
\title{nearestValue
Return the nearest value in data}
\usage{
nearestValue(value, data)

nearestValue(value, data)
}
\arguments{
\item{value}{-> reference value}

\item{data}{-> array of sort values}
}
\value{
nearest value in data to value; -1 if value es out of range

nearest value in data to value; -1 if value es out of range
}
\description{
nearestValue
Return the nearest value in data

nearestValue
Return the nearest value in data
}
