% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{peaksDeconvolved}
\alias{peaksDeconvolved}
\title{peaksDeconvolved
Para cada pico de magnitud compuesto que contiene a más de un pico simples de sirem,
se anotan los picos de sirem incluidos y las gausianas asociadas}
\usage{
peaksDeconvolved(peaksInfo, gaussInfo, uPeakList = c())
}
\arguments{
\item{peaksInfo}{-> sirem peaks, from rGetSiremPeaks()}

\item{gaussInfo}{-> gaussians,   from rGetGaussians()}

\item{uPeakList}{-> lista de picos compuestos a considerar, esten o no deconvolucionados}
}
\value{
a list: siremMassMatrix, gaussMassMatrix, gaussMagMatrix, gaussMassList
siremMassMatrix mantiene las masas asociadas a los picos de sirem
gaussMassMatrix mantiene las masas asociadas a las gausianas luego del ajuste con el valor promediado de magnitudes
gaussMagMatrix  mantiene las magnitudes asociadas a cada gausiana
gaussList es un array con las gausianas de gaussMassMatrix
siremMassMatrix & gaussMassMatrix and gaussMagMatrix have the same format:
rows=united magPeaks deconvolved;
col[] =central mass of each peak
col\link{1}=index to magnitude united Peak
col\link{2}=low  mass to magnitude united peak
col\link{3}=high mass to magnitude united peak
col\link{4}=number of single peaks into magnitude peak
col\link{5...}=list of masses into magnitude united peaks

#attn: en peakInfo, si en un pico de  magnitud no se encuentra ningún máximo de sirem,
}
\description{
peaksDeconvolved
Para cada pico de magnitud compuesto que contiene a más de un pico simples de sirem,
se anotan los picos de sirem incluidos y las gausianas asociadas
}
