% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGetGaussians.R
\name{rGetGaussians}
\alias{rGetGaussians}
\title{rGetGaussians
Gets the Gaussians that decompose the given concentration data.
Makes use of the peak information returned by rGetSiremPeaks().}
\usage{
rGetGaussians(rMSIData, siremInfo, pixel, minMeanValue)
}
\arguments{
\item{rMSIData}{-> sample data obtained from the file with rMSI2::LoadMsiData().}

\item{siremInfo}{-> sirem data and peaks obtained from rGetSiremPeaks().}

\item{pixel}{-> pixel whose partial spectrum is deconvolved. if pixel<=0, averaged data is used}

\item{minMeanValue}{-> minimum averaged value of a peak to consider it valid.}
}
\value{
a list:
gaussians -> matrix with the parameters of the Gaussians.
xAxis -> X axis (Daltons) (if error xAxis=0)
yAxis -> Y axis (concentration) (if error yAxis=0)
If pixel<=0 the averaged values are returned.
If pixel > 0, the part of the spectrum associated with the X axis is returned
}
\description{
rGetGaussians
Gets the Gaussians that decompose the given concentration data.
Makes use of the peak information returned by rGetSiremPeaks().
}
