% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGetSirem.R
\name{rGetSirem}
\alias{rGetSirem}
\title{rGetSirem
Get information about sirem and its associated spikes.}
\usage{
rGetSirem(rMSIData, params, initMass, finalMass)
}
\arguments{
\item{rMSIData}{-> sample data obtained from the file with rMSI2::LoadMsiData().}

\item{params}{-> parameters for sirem and for peaks.}

\item{initMass}{-> mass, in Daltons, corresponding to the first image.}

\item{finalMass}{-> mass, in Daltons, corresponding to the last image.}
}
\value{
a list: sirem->sirem information; scanMean->averaged values of each image; sirem->sirem value of each image.
}
\description{
rGetSirem
Get information about sirem and its associated spikes.
}
