% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGetSiremPeaks.R
\name{rGetSiremPeaks}
\alias{rGetSiremPeaks}
\title{rGetSiremPeaks:
get information about Sirem and its associated spikes.}
\usage{
rGetSiremPeaks(rMSIData, params, initMass, finalMass)
}
\arguments{
\item{rMSIData}{Sample data obtained from the file with rMSI2::LoadMsiData().}

\item{params}{Parameters for sirem and for peaks.
algorithm         -> 0 = sirem; 1=entropy.
cutLeves          -> cut levels to apply to each image to generate binary images (percentiles).
magSensitivity    -> variable sensitivity depending on the concentration level\link{0:maxConcentration} (bounded between limits).
siremSensitivity  -> variable sensitivity depending on the sirem level\link{0:1} (bounded between limits).
minMeanPxMag      -> minimum averaged concentration value to be considered.
minSectionDensity -> minimum active pixel density value in the image to be considered.
noiseLevel        -> absolute noise Level\link{0:maxConcentration}: lower values are considered null.
pxCoord_X         -> X coordinates of each pixel.
pxCoord_Y         -> Y coordinates of each pixel.
referenceType     -> reference image type:
0=obtained from concentration info (data)
1=the image with the highest average value is adopted
2=all pixels have significant concentration.
scanReference     -> reference image to determine the area to consider.
tileSide          -> number of side pixels of the square tile: 1(1x1), 2(2x2), 3(3x3), 4(4x4)
normalization     -> 0=none
-> 1=TIC
-> 2=RMS
-> 3=MAX}

\item{initMass}{Mass, in Daltons, corresponding to the first image to load.}

\item{finalMass}{Mass, in Daltons, corresponding to the last  image to load.}
}
\value{
A list:
siremPeaks -> sirem and peak information;
massAxis   -> mass axis corresponding to the images.
}
\description{
rGetSiremPeaks:
get information about Sirem and its associated spikes.
}
