% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGetSiremPeaks.R
\name{rLoadImages}
\alias{rLoadImages}
\title{rLoadImages
Returns an array with a set of images.}
\usage{
rLoadImages(rMSIData, initMass, size)
}
\arguments{
\item{rMSIData}{-> sample data obtained from the file with rMSI2::LoadMsiData().}

\item{initMass}{-> mass in Daltons corresponding to the first image.}

\item{size}{-> number of images.}
}
\value{
a matrix: column is the image corresponding to consecutive masses; rows contains the concentration data for each pixel.
}
\description{
rLoadImages
Returns an array with a set of images.
}
