% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{rMSI2PeaksFilterDeconv}
\alias{rMSI2PeaksFilterDeconv}
\title{siremPeaksFilterDeconv
Retorna una matriz con los centroides de rMSI2 contenidos en los picos compuestos deconvolucionados de rSIREM}
\usage{
rMSI2PeaksFilterDeconv(siremPeaks, gaussInfo, rMSIPeaks)
}
\arguments{
\item{siremPeaks}{-> sirem peaks, from rGetSiremPeaks()}

\item{rMSIPeaks}{-> rMSI2 peaks, from rMSI2::LoadPeakMatrix() from rMSI2::processWizard()}
}
\value{
una matriz con los centroides de rMSI2 contenidos en los picos compuestos de rSIREM
columna 1   referencia al pico compuesto de rSIREM
columna 2.. centroides de rMSI2
}
\description{
siremPeaksFilterDeconv
Retorna una matriz con los centroides de rMSI2 contenidos en los picos compuestos deconvolucionados de rSIREM
}
