% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotConcentration}
\alias{rPlotConcentration}
\title{rPlotConcentration
Presents an image with peaks concentration information.
In red the magnitude curve.
Blue circles indicate data.}
\usage{
rPlotConcentration(rMSIData, pixel, minMz, maxMz)
}
\arguments{
\item{peakInfo:}{list returned from rGetSiremPeaks().}

\item{minMz:}{minimun mz}

\item{maxMz:}{maximun mz}
}
\description{
rPlotConcentration
Presents an image with peaks concentration information.
In red the magnitude curve.
Blue circles indicate data.
}
