% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotDeconv}
\alias{rPlotDeconv}
\title{rPlotDeconv
Presents an image with the Gaussians passed as arguments,
along with the curve resulting from the sum of all of them (in red).
The blue circles indicate the original concentration data to be adjusted.}
\usage{
rPlotDeconv(drawInfo)
}
\arguments{
\item{drawInfo$gaussians:}{column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.}

\item{drawInfo$axis:}{column 1 -> Y axis (concentration data to adjust).
column 2 -> X axis (in Daltons).}
}
\value{
-1 if KO; 0 if OK
}
\description{
rPlotDeconv
Presents an image with the Gaussians passed as arguments,
along with the curve resulting from the sum of all of them (in red).
The blue circles indicate the original concentration data to be adjusted.
}
