% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotDeconv3}
\alias{rPlotDeconv3}
\title{rPlotDeconv3
Presents an image with the Gaussians of tree peaks,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original concentration data to be adjusted.
The parameters come from rGetGaussians()}
\usage{
rPlotDeconv3(
  drawInfo1,
  drawInfo2,
  drawInfo3,
  rMSI2_peaks1 = -1,
  rMSI2_peaks2 = -1,
  rMSI2_peaks3 = -1
)
}
\arguments{
\item{drawInfo1:}{gaussians info from low resolution peak}

\item{drawInfo2:}{gaussians info from median resolution peak}

\item{drawInfo3:}{gaussians info from high resolution peak
drawInfo$gaussians:
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.
drawInfo$axis:
column 1 -> Y axis (concentration data to adjust).
column 2 -> X axis (in Daltons).}
}
\value{
-1 if KO; 0 if OK
}
\description{
rPlotDeconv3
Presents an image with the Gaussians of tree peaks,
along with the curve resulting from the sum of them (in red).
The blue circles indicate the original concentration data to be adjusted.
The parameters come from rGetGaussians()
}
