% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotPeaks}
\alias{rPlotPeaks}
\title{rPlotPeaks
Presents an image with the centroids of two data set into a mass range,
The parameters come from rGetGaussians() and rMSI2::LoadPeakMatrix()}
\usage{
rPlotPeaks(siremPeak, MSI2peak, minMass, maxMass)
}
\arguments{
\item{siremPeak}{-> gaussians info matrix from return of rGetGaussians()}

\item{MSI2peak}{-> list from rMSI2::LoadPeakMatrix()}

\item{minMass}{-> high mass to analyze (Da)
siremPeak$gaussians:
column 1 -> mean values of each Gaussian.
column 2 -> standard deviation.
column 3 -> factor associated with each Gaussian.
MSI2peak$mass: mass vector
MSI2peak$intensity:
columns -> concentration for each mass.
rows    -> concentration for each pixel.}
}
\value{
-1 if KO; 0 if OK
}
\description{
rPlotPeaks
Presents an image with the centroids of two data set into a mass range,
The parameters come from rGetGaussians() and rMSI2::LoadPeakMatrix()
}
