% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rPlotDeconv.R
\name{rPlotSirem}
\alias{rPlotSirem}
\title{rPlotSirem
Presents an image with peak concentration and sirem information.
In red the magnitude curve and in green the sirem curve.
Blue circles indicate data.}
\usage{
rPlotSirem(peaksInfo, minMz, maxMz)
}
\arguments{
\item{peakInfo:}{list returned from rGetSiremPeaks().}

\item{minMz:}{minimun mz}

\item{maxMz:}{maximun mz}
}
\description{
rPlotSirem
Presents an image with peak concentration and sirem information.
In red the magnitude curve and in green the sirem curve.
Blue circles indicate data.
}
