% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{scans2Daltons}
\alias{scans2Daltons}
\title{scans2Daltons
convierte un array de scans a Daltons}
\usage{
scans2Daltons(scans, massAxis)
}
\arguments{
\item{scans}{-> array de valores correspondientes a m/z en massAxis; scans\link{1}=0...
ejemplo: scans=5.5 -> massAxis\link{5} + (massAxis\link{6}-massAxis\link{5})/2}

\item{massAxis}{-> eje de masas}
}
\value{
array en Daltons correspondientes al array en scans
}
\description{
scans2Daltons
convierte un array de scans a Daltons
}
