% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{sirem_vs_rMSI2}
\alias{sirem_vs_rMSI2}
\title{sirem_vs_rMSI2
genera datos estadísticos con las desviaciones de ambos algoritmos respecto a una muestra patrón}
\usage{
sirem_vs_rMSI2(sample, SNR)
}
\arguments{
\item{sample}{-> sample. Valids: "C30k" y "C60k"}

\item{SNR}{-> signal to noise ratio. Valids: 1, 2, 3, 5, 7}
}
\value{
nothing
}
\description{
sirem_vs_rMSI2
genera datos estadísticos con las desviaciones de ambos algoritmos respecto a una muestra patrón
}
