% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_old.R
\name{siremPeaksDeconvolved}
\alias{siremPeaksDeconvolved}
\title{siremPeaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados
Los resultados son previos al ajuste por gausianas}
\usage{
siremPeaksDeconvolved(peaksInfo)
}
\arguments{
\item{peaksInfo}{-> list returned from rGetSiremPeaks()}
}
\value{
a matrix: rows=magPeaks deconvolved;
col[] =mass of maximun
col\link{1}=magitude Peak
col\link{2}=low  sirem peak
col\link{3}=high sirem peak
col\link{4}=number of peaks into magnitude peak
}
\description{
siremPeaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados
Los resultados son previos al ajuste por gausianas
}
