% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{siremPeaksFilter}
\alias{siremPeaksFilter}
\title{siremPeaksFilter
Retorna un array indicando si el pico compuesto incluye a picos de rMSI2
Sólo se consideran aquellos que contienen a un pico de rMSI}
\usage{
siremPeaksFilter(siremPeaks, rMSIPeaks)
}
\arguments{
\item{siremPeaks}{-> sirem peaks, from rGetSiremPeaks()}

\item{rMSIPeaks}{-> rMSI2 peaks, from  rMSI2::LoadPeakMatrix() from rMSI2::processWizard()}
}
\value{
array indicando si el pico compuesto incluye a picos de rMSI2
}
\description{
siremPeaksFilter
Retorna un array indicando si el pico compuesto incluye a picos de rMSI2
Sólo se consideran aquellos que contienen a un pico de rMSI
}
