% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{violinSirem}
\alias{violinSirem}
\title{violinSirem()
imagen violin para las desviaciones de rSIREM y rMSI2}
\usage{
violinSirem(type = "SIREM")
}
\arguments{
\item{type:}{SIREM, SIREM_d, SIREM_nd, MSI2, MSI2_MSI2, MSI2_SIREM
SIREM    -> para todos los centroides de rSIREM
SIREM_d  -> para los controides exclusivos de rSIREM (no existentes en rMSI2)
SIREM_nd -> para los centroides compartidos entre rSIREM y rMSI2
MSI2     -> para todos los centroides de rMSI2
rMSI2_MSI2  -> rMSI2 sobre los centroides comunes entre rMSI2 y rSIREM
rMSI2_SIREM -> SIREM sobre los centroides comunes entre rMSI2 y rSIREM}
}
\value{
a list with two statistical data matrices.
}
\description{
violinSirem()
imagen violin para las desviaciones de rSIREM y rMSI2
}
