% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{fitQualitySirem}
\alias{fitQualitySirem}
\title{fitQualitySiremDeconv()
Segrega los picos compuestos en los que existe deconvolución, indica los picos simples que lo componen
e identifica el pico de alta resolución con el que tiene el mejor ajuste}
\usage{
fitQualitySirem(
  reference,
  testSirem,
  testGauss,
  refMinMag = 1e-06,
  testMinMag = 1e-06,
  minMass = 0,
  maxMass = 0
)
}
\arguments{
\item{reference}{-> gaussians info from rGetGaussians() (high resolution)}

\item{testSirem}{-> sirem     info from rGetGaussians() (low resolution)}

\item{testGauss}{-> gaussians info from rGetGaussians() (low resolution)}

\item{refMinMag}{-> mínima magnitud de picos para su consideración (high resolution)}

\item{testMinMag}{-> mínima magnitud de picos para su consideración (low resolution)}

\item{minMass}{-> high mass to analyze (Da)}
}
\value{
a matrix with columns: "mzTest", "mzRef", "ppm", "maxDev", "repe", "deconv"
mzTest -> mz from low  resolution sample
mzRef  -> mz from high resolution sample
ppm    -> deviation from low to high mz
maxDev -> '1' if greater than ppm of one scan
repe   -> '1' if mzRef is the same for two consecutives mzTest
deconv -> '1' if is a deconvolved peak
}
\description{
fitQualitySirem()
Identifica el pico de alta resolución más próximo al pico de baja resolución e indica la desviación
Lo hace sobre todos los picos. Para info de Sirem
}
\details{
fitQualityPere()
Identifica el pico de alta resolución más próximo al pico de baja resolución e indica la desviación
Lo hace sobre todos los picos. Para info de rMSI2 (Pere)

peaksDeconvolved()
Para cada pico de magnitud compuesto que contiene a picos simples de sirem,
se anotan los picos de sirem incluidos.

getUpeakSiremMassesFromUpeak()
retorna un array con las masas de sirem asociadas a un pico de magnitud

getUpeakGaussFromMass()
retorna un array con las masas de sirem contenidas en el mismo pico compuesto que la masa dada

siremPeaksFilter()
Retorna un array indicando si el pico compuesto debe ser considerado o no
Sólo se consideran aquellos que contienen a un pico de rMSI2

sirem_vs_rMSI2()
genera datos estadísticos con las desviaciones de ambos algoritmos respecto a una muestra patrón
fitQualitySirem
Identifica el pico de alta resolución más próximo al pico de baja resolución e indica la desviación
Lo hace sobre todos los picos. Para info de Sirem
}
