% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{peaksDeconvolved}
\alias{peaksDeconvolved}
\title{peaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados
Los resultados son previos al ajuste por gausianas}
\usage{
peaksDeconvolved(peaksInfo, gaussInfo)
}
\arguments{
\item{peaksInfo}{-> list returned from rGetSiremPeaks()}
}
\value{
a matrix: rows=magPeaks deconvolved;
col[] =mass of maximun
col\link{1}=index to magnitude united Peak
col\link{2}=low  mass to magnitude united peak
col\link{3}=high mass to magnitude united peak
col\link{4}=number of peaks into magnitude peak
col\link{5...}=list of masses into magnitude united peaks
}
\description{
peaksDeconvolved
identifica los picos de magnitud que han sido deconvolucionados
Los resultados son previos al ajuste por gausianas
}
