% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rGetSiremPeaks.R
\name{rGetIndexFromMass}
\alias{rGetIndexFromMass}
\title{rGetIndexFromMass
Returns the index to the full mass axis whose mass is closest to the given mass.
The algorithm of successive approximations is used.}
\usage{
rGetIndexFromMass(rMSIData, mass)
}
\arguments{
\item{rMSIData}{-> sample data obtained from the file with rMSI2::LoadMsiData().}

\item{mass}{-> mass in Daltons}
}
\value{
The index to mass if OK; -1 if KO
}
\description{
rGetIndexFromMass
Returns the index to the full mass axis whose mass is closest to the given mass.
The algorithm of successive approximations is used.
}
